\name{deviance.KFS}
\alias{deviance.KFS}
\title{Deviance of a State Space Model}
\usage{
\method{deviance}{KFS}(object, ...)
}
\arguments{
  \item{object}{An object of class \code{KFS}.}

  \item{\dots}{Ignored.}
}
\value{
The value of the deviance extracted from object.
}
\description{
Returns the deviance of a object of class \code{KFS}.
}

