\name{GlobalTemp}
\docType{data}
\alias{GlobalTemp}
\title{Two series of average global temperature deviations for years 1880-1987}
     \description{
       This data set contains two series of average global temperature deviations for years 1880-1987. These series are same as used in Shumway and Stoffer (2006), where they are known as HL and Folland series. For more details, see Shumway and Stoffer (2006, p. 327).
     }
     \usage{GlobalTemp}
     \format{An 2 times 108 array containing 216 observations.}
     \source{http://lib.stat.cmu.edu/general/stoffer/tsa2/}
     \references{
Shumway, Robert H. and Stoffer, David S. (2006). Time Series Analysis and Its Applications: With R examples. 
     }
     \keyword{datasets}


