% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KFAS-defunct.R
\name{KFAS-defunct}
\alias{KFAS-defunct}
\alias{deviance.KFS}
\alias{subset.SSModel}
\alias{subset<-}
\alias{subset<-.SSModel}
\title{Defunct Functions of Package KFAS}
\usage{
\method{deviance}{KFS}(object, ...)

\method{subset}{SSModel}(x, element, states, etas, series, times, ...) <- value

subset(x, ...) <- value

\method{subset}{SSModel}(x, element, states, etas, series, times, ...)
}
\description{
The function listed here are removed from KFAS.
}
\details{
Subset-based methods were removed as they were somewhat confusingly
named as the \code{subset} generic in \code{base}, and most likely not
that useful (compared to \code{\link{[<-.SSModel}}).

Deviance.KFS was removed as it was mostly useless. The value was not a \eqn{-2*(logL-logL*)}
where \eqn{L} is the likelihood and \eqn{L*} is the saturated likelihood.
Instead it was based on the conditional likelihood \eqn{p(y|theta)} i.e. it disregards
the effect of hidden states. Therefore the value
returned by this function did not make much sense in non-GLM setting.
}
\keyword{internal}

