\name{plot.approxSSM}
\alias{plot.approxSSM}
\title{Plot Approximating Gaussian State Space Model}
\usage{
  \method{plot}{approxSSM} (x, y = NULL,
    legend.position = "bottomright", ...)
}
\arguments{
  \item{x}{approxSSM object}

  \item{y}{Ignored.}

  \item{legend.position}{Position of the legend as a
  keyword. See function legend for details.}

  \item{...}{Graphical parameters for function plot.ts.}
}
\description{
  Plot Approximating Gaussian State Space Model
}

