\name{print.KFS}
\alias{print.KFS}
\title{Print Ouput of Kalman Filter and Smoother}
\usage{
\method{print}{KFS}(x, digits = max(3L, getOption("digits") - 3L), ...)
}
\arguments{
  \item{x}{output object from function KFS.}

  \item{digits}{minimum number of digits to be printed.}

  \item{...}{Ignored.}
}
\description{
Print Ouput of Kalman Filter and Smoother
}

