% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KFAS-package.R
\docType{data}
\name{alcohol}
\alias{alcohol}
\title{Alcohol related deaths in Finland 1969--2013}
\format{A multivariate time series object with 45 times 8 observations.}
\source{
Statistics Finland \url{http://pxnet2.stat.fi/PXWeb/pxweb/en/StatFin/}.
}
\description{
A multivariate time series object containing the number of alcohol related
deaths and population sizes (divided by 100000) of Finland in four age groups.
}
\keyword{datasets}

