% Generated by roxygen2 (4.0.2): do not edit by hand
\name{artransform}
\alias{artransform}
\title{Mapping real valued parameters to stationary region}
\usage{
artransform(param)
}
\arguments{
\item{param}{Real valued parameters for the transformation.}
}
\value{
transformed The parameters satisfying the stationary constrains.
}
\description{
Function \code{artransform} transforms \eqn{p} real valued parameters to
stationary region of \eqn{p}th order autoregressive process using
parametrization suggested by Jones (1980). Fortran code is a converted from
stats package's C-function partrans.
}

