\docType{data}
\name{boat}
\alias{boat}
\title{Oxford-Cambridge boat race results 1829-2000}
\format{A time series object containing 172 observations.}
\source{
http://www.ssfpack.com/DKbook.html
}
\description{
Results of the annual boat race between universities of
Oxford (0) and Cambridge (1).
}
\references{
Koopman, S.J. and Durbin J. (2001).  Time Series Analysis
by State Space Methods. Oxford: Oxford University Press.
}
\keyword{datasets}

