% Generated by roxygen2 (4.0.2): do not edit by hand
\name{deviance.KFS}
\alias{deviance.KFS}
\title{Deviance of a State Space Model}
\usage{
\method{deviance}{KFS}(object, ...)
}
\arguments{
\item{object}{An object of class \code{KFS}.}

\item{\dots}{Ignored.}
}
\value{
The value of the deviance extracted from object.
}
\description{
Returns the deviance of a object of class \code{KFS} defined as the sum of squared deviance residuals.
}
\details{
Note that this is NOT a $-2*(logL-logL*)$ where $L$ is the likelihood and $L*$ is the saturated likelihood.
Instead this is based on the conditional likelihood $p(y|theta)$ i.e. it disregards the effect of hidden states.
Therefore the value returned by this function might not make sense for non-GLM setting.
it is just
Thus this deviance is defined as -2*
}

