% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deviance.KFS.R
\name{deviance.KFS}
\alias{deviance.KFS}
\title{Deviance of a State Space Model}
\usage{
\method{deviance}{KFS}(object, ...)
}
\arguments{
\item{object}{An object of class \code{KFS}.}

\item{\dots}{Ignored.}
}
\value{
The value of the deviance extracted from object.
}
\description{
Returns the deviance of an object of class \code{KFS} defined as the sum of squared deviance
residuals.
}
\details{
This function is deprecated. Note that this is NOT a \eqn{-2*(logL-logL*)} where \eqn{L} is the
likelihood and \eqn{L*} is the saturated likelihood. Instead this is based on the conditional
likelihood \eqn{p(y|theta)} i.e. it disregards the effect of hidden states. Therefore the value
returned by this function might not make sense for non-GLM setting.
}

