% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/extract.SSModel.R, R/subset.SSModel.R
\name{[<-.SSModel}
\alias{[.SSModel}
\alias{[<-.SSModel}
\alias{subset.SSModel}
\alias{subset<-}
\alias{subset<-.SSModel}
\title{Extract or Replace Parts of a State Space Model}
\usage{
\method{[}{SSModel}(x, element, states, etas, series, times, ...) <- value

\method{[}{SSModel}(x, element, states, etas, series, times, drop = TRUE, ...)

\method{subset}{SSModel}(x, element, states, etas, series, times, ...) <- value

subset(x, ...) <- value

\method{subset}{SSModel}(x, element, states, etas, series, times, ...)
}
\arguments{
\item{x}{Object of class \code{SSModel}.}

\item{element}{Which element is chosen. Possible choices are
'y','Z','H','T','R','Q','a1','P1','P1inf', and 'u'.}

\item{states}{Which states are chosen. Either a numeric vector containing
the indices of the states, or a character vector defining the
types of the states.
Possible choices are \dQuote{all}, \dQuote{arima}, \dQuote{custom},
\dQuote{cycle}, \dQuote{seasonal}, \dQuote{trend}, or \dQuote{regression}.
These can be combined. Default is \dQuote{all}.}

\item{etas}{Which disturbances eta are chosen. Used for elements \dQuote{R}
and \dQuote{Q}. Either a numeric vector containing the indices of the
etas, or a character vector defining the types of the etas.
Possible choices are \dQuote{all}, \dQuote{arima}, \dQuote{custom},
\dQuote{cycle}, \dQuote{seasonal}, \dQuote{trend}, or \dQuote{regression}.
These can be combined.}

\item{series}{Numeric. Which series are chosen. Used for elements
\dQuote{y}, \dQuote{Z}, and \dQuote{u}.}

\item{times}{Numeric. Which time points are chosen.}

\item{...}{ignored.}

\item{value}{A value to be assigned to x.}

\item{drop}{Logical. If TRUE the result is coerced to the lowest possible
dimension. Default is FALSE (which differs from default R behaviour).}
}
\value{
A selected subset of the chosen element or a value.
}
\description{
S3 methods for getting and setting parts of object of class
\code{SSModel}. These methods ensure that dimensions of system matrices are
not altered.
}
\details{
Method for \code{subset} and corresponding replacement
method are deprecated.
}
\examples{
set.seed(1)
model <- SSModel(rnorm(10) ~ 1)
model['H']
model['H'] <- 10
# H is still an array:
model['H']
logLik(model)
model$H <- 1
# model['H'] throws an error as H is now scalar:
model$H
logLik(model, check.model = TRUE) #with check.model = FALSE R crashes!
}

