% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitted.KFS.R
\name{fitted.KFS}
\alias{fitted.KFS}
\title{Extract Fitted Values of State Space Model}
\usage{
\method{fitted}{KFS}(object, start = NULL, end = NULL, filtered = FALSE,
  ...)
}
\arguments{
\item{object}{An object of class \code{KFS}.}

\item{start}{The start time of the period of interest. Defaults to first time
point of the object.}

\item{end}{The end time of the period of interest. Defaults to the last time
point of the object.}

\item{filtered}{Logical, return filtered instead of smoothed estimates of
state vector. Default is \code{FALSE}.}

\item{...}{Ignored.}
}
\value{
Multivariate time series containing fitted values.
}
\description{
Extracts fitted values from output of \code{KFS}, i.e. one-step ahead
predictions  \eqn{f(θ_t | y_{t-1}, \ldots, y_1)}{
f(θ[t] | y[t-1], ... , y[1]),} (\code{m}) or smoothed estimates
\eqn{f(θ_t | y_n, \ldots, y_1)}{f(θ[t] | y[n], ... , y[1]),} (\code{muhat}),
where \eqn{f} is the inverse of the link function
(identity in Gaussian case), except in case of Poisson distribution where
\eqn{f} is multiplied with the exposure \eqn{u_t}{u[t]}.
}
\examples{
data("sexratio")
model <- SSModel(Male ~ SSMtrend(1,Q = list(NA)),u = sexratio[, "Total"],
  data = sexratio, distribution = "binomial")
model <- fitSSM(model,inits = -15, method = "BFGS")$model
out <- KFS(model)
identical(drop(out$muhat), fitted(out))
}
\seealso{
\code{\link{signal}} for partial signals and their covariances.
}

