% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.objects.R
\name{print.KFS}
\alias{print.KFS}
\title{Print Ouput of Kalman Filter and Smoother}
\usage{
\method{print}{KFS}(x, type = "state", digits = max(3L,
  getOption("digits") - 3L), ...)
}
\arguments{
\item{x}{output object from function KFS.}

\item{type}{What to print. Possible values are \code{"state"} (default),
\code{"signal"}, and  \code{"mean"}. Multiple choices are allowed.}

\item{digits}{minimum number of digits to be printed.}

\item{...}{Ignored.}
}
\description{
Print Ouput of Kalman Filter and Smoother
}
