% Generated by roxygen2 (4.0.2): do not edit by hand
\name{rstandard.KFS}
\alias{rstandard.KFS}
\title{Extract Standardized Residuals from KFS output}
\usage{
\method{rstandard}{KFS}(model, type = c("recursive", "observation", "state",
  "deviance", "pearson"), ...)
}
\arguments{
\item{model}{KFS object}

\item{type}{Type of residuals. See details.}

\item{...}{Ignored.}
}
\description{
Extract Standardized Residuals from KFS output
}
\details{
For object of class KFS with fully Gaussian observations, several
  types of standardized residuals can be computed:

  \itemize{
  \item "recursive": For Gaussian models the one-step ahead prediction
  residuals defined as
  \deqn{v_{t,i})/\sqrt{F_{i,t}},}{v[t,i])/\sqrt{F[i,t]},} with residuals
  being undefined in diffuse phase. Note that even in multivariate case these
  standardized residuals coincide with the ones obtained from the Kalman
  filter without the sequential processing (which is not true for the
  non-standardized innovations).
  For non-Gaussian models the recursive residuals are obtained as
  \deqn{L^{-1}_t (y_{t}-\mu_{t}),}{L^(-1)[t](y[t]-\mu[t]),} where
  \eqn{L_t}{L[t]} is the lower triangular matrix from Cholesky decomposition
  of \eqn{V(y_t)+V(\mu_t)}{V(y[t])+V(\mu[t])}, and both the variance function \eqn{V(y_t)}{V(y[t])} and
  \eqn{V(\mu_t)}{V(\mu[t])} are based on the filtered means \eqn{\mu_t}{\mu[t]}$. Computing these for large
  non-Gaussian models can be time consuming as filtering is needed.

  \item "observation":  Residuals based on the smoothed observation
  disturbance terms \eqn{\epsilon} are defined as
  \deqn{L^{-1}_t \hat\epsilon_t, \quad t=1,\ldots,n,}{L^{-1}[t] \epsilon[t], t=1,\ldots,n,}
  where \eqn{L_t}{L[t]} is the lower triangular matrix from Cholesky
  decomposition of \eqn{V_{\epsilon,t}}{V[\epsilon,t]}.


  \item "state":  Residuals based on the smoothed state disturbance terms
  \eqn{\eta} are defined as \deqn{L^{-1}_t \hat \eta_t, \quad
  t=1,\ldots,n,}{L^{-1}[t] \hat \eta[t], t=1,\ldots,n,} where \eqn{L_t}{L[t]} is
  the lower triangular matrix from Cholesky decomposition of
  \eqn{V_{\eta,t}}{V[\eta,t]}.


  \item 'pearson':  Standardized Pearson residuals
  \deqn{L^{-1}_t(y_{t}-\theta_{i}), \quad t=1,\ldots,n,}{L^(-1)[t](y[t]-\theta[t]), t=1,\ldots,n,} where
  \eqn{L_t}{L[t]} is the lower triangular matrix from Cholesky decomposition
  of \eqn{V(y_t)-V(\mu_t)}{V(y[t])-V(\mu[t])}. For gaussian models, these coincide with the
  standardized smoothed \eqn{\epsilon} disturbance residuals, and for
  generalized linear models these coincide with the standardized Pearson
  residuals (hence the name).


\item 'deviance': Deviance residuals.
Deprecated. This option was meant to be used only for the GLM comparisons, as their generalization to other models is lacking,
but these will be completely removed in future in order to avoid misleading results in non-GLM settings.
}
}

